:truck: Move progress visualization to dedicated folder
